Refactor code to be more maintainable
