+++ src/lib.rs
@@ -62,7 +62,10 @@ println!("Data length: {:?}", image.data().len());
 */
 #![warn(missing_docs)]
 
-use std::{borrow::Cow, fmt, fs, io, panic, path};
+use std::{fs, panic, path};
+
+pub use error::{ConfigError, DecodingError};
+pub use image::{BitDepth, ColorSpace, ImageData, OutputFormat};
 
 /// Decoders for images
 #[deprecated(since = "0.2.0", note = "use the Decoder struct instead")]
@@ -71,7 +74,13 @@ pub mod decoders;
 #[deprecated(since = "0.2.0", note = "use the Encoder struct instead")]
 pub mod encoders;
 
-/// Config from command line input
+/// All errors that can occur
+pub mod error;
+
+/// Image data
+pub mod image;
+
+/// Config for encoder
 #[derive(Debug)]
 pub struct Config<'a> {
     input: &'a [path::PathBuf],
@@ -82,18 +91,26 @@ pub struct Config<'a> {
 impl<'a> Config<'a> {
     /// Builds config from parameters
     ///
-    /// # Result
+    /// # Examples
     ///
-    /// - [`Config`] if Ok
-    /// - [`ConfigError`] if
-    ///   - Quality under 0 or greater than 1
-    ///   - input is empty
+    /// ```
+    /// # use rimage::{Config, OutputFormat};
+    /// # use std::path;
+    /// let input = &[path::PathBuf::from("tests/files/basi0g01.jpg")];
+    /// let quality = 100.0;
+    /// let output_format = OutputFormat::MozJpeg;
+    /// let config = Config::build(input, quality, output_format).unwrap();
+    /// ```
+    ///
+    /// # Errors
+    /// - [`ConfigError::QualityOutOfBounds`] if quality is not in range 0.0 - 100.0
+    /// - [`ConfigError::InputIsEmpty`] if input is empty
     pub fn build(
         input: &'a [path::PathBuf],
         quality: f32,
         output_format: OutputFormat,
     ) -> Result<Self, ConfigError> {
-        if quality < 0.0 || quality > 1.0 {
+        if quality < 0.0 || quality > 100.0 {
             return Err(ConfigError::QualityOutOfBounds);
         }
 
@@ -107,14 +124,17 @@ impl<'a> Config<'a> {
             output_format,
         })
     }
+    #[inline]
     /// Gets input array of paths from config
     pub fn input(&self) -> &[path::PathBuf] {
         &self.input
     }
+    #[inline]
     /// Gets quality of output images from config
     pub fn quality(&self) -> f32 {
         self.quality
     }
+    #[inline]
     /// Gets format of output images from config
     pub fn output_format(&self) -> &OutputFormat {
         &self.output_format
@@ -122,278 +142,20 @@ impl<'a> Config<'a> {
 }
 
 impl<'a> Default for Config<'a> {
+    #[inline]
     fn default() -> Self {
         Self {
             input: &[],
-            quality: 0.75,
+            quality: 75.0,
             output_format: OutputFormat::MozJpeg,
         }
     }
-}
-
-/// Image data from decoder
-///
-/// # Examples
-///
-/// ```
-/// # use rimage::{Decoder, DecodingError};
-/// # use std::path;
-/// # let path = path::PathBuf::from("tests/files/basi0g01.jpg");
-/// let d = Decoder::build(&path)?;
-///
-/// let image = d.decode()?;
-///
-/// // Get something from image data
-/// println!("Color Space: {:?}", image.color_space());
-/// println!("Bit Depth: {:?}", image.bit_depth());
-/// println!("Size: {:?}", image.size());
-/// println!("Data length: {:?}", image.data().len());
-/// # Ok::<(), DecodingError>(())
-/// ```
-pub struct ImageData<'a> {
-    color_space: ColorSpace,
-    bit_depth: BitDepth,
-    width: usize,
-    height: usize,
-    data: Cow<'a, [u8]>,
 }
 
 /// Decoder used to get image data from file
 pub struct Decoder<'a> {
     path: &'a path::PathBuf,
     raw_data: Vec<u8>,
-}
-
-// Write encoder struct
-/// Encoder used to encode image data to file
-pub struct Encoder<'a> {
-    path: &'a path::PathBuf,
-    image_data: ImageData<'a>,
-    quality: f32,
-}
-
-/// Supported output format
-///
-/// # Examples
-/// ```
-/// # use rimage::OutputFormat;
-/// # use std::str::FromStr;
-/// let format = OutputFormat::from_str("mozjpeg").unwrap();
-/// println!("Format: {}", format);
-/// ```
-#[derive(Debug, Clone, Copy)]
-pub enum OutputFormat {
-    /// MozJpeg image
-    MozJpeg,
-    /// Browser Png image
-    Png,
-    /// OxiPng image
-    Oxipng,
-}
-
-impl std::str::FromStr for OutputFormat {
-    type Err = Cow<'static, str>;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "mozjpeg" | "jpg" | "jpeg" => Ok(OutputFormat::MozJpeg),
-            "png" => Ok(OutputFormat::Png),
-            "oxipng" => Ok(OutputFormat::Oxipng),
-            _ => Err(format!("{} is not a valid output format", s).into()),
-        }
-    }
-}
-
-impl fmt::Display for OutputFormat {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            OutputFormat::MozJpeg => write!(f, "jpeg"),
-            OutputFormat::Png => write!(f, "png"),
-            OutputFormat::Oxipng => write!(f, "oxipng"),
-        }
-    }
-}
-
-/// An error that occurred during decoding a image
-///
-/// # Examples
-/// ```
-/// # use rimage::{Decoder, DecodingError};
-/// # use std::path;
-/// # let path = path::PathBuf::from("tests/files/basi0g01.jpg");
-/// let d = Decoder::build(&path)?;
-/// let image = d.decode();
-/// match image {
-///     Ok(_) => println!("Image decoded"),
-///     Err(e) => println!("Error: {}", e),
-/// }
-/// # Ok::<(), DecodingError>(())
-/// ```
-#[derive(Debug)]
-pub enum DecodingError {
-    /// A [`io::Error`] if file failed to read, find, etc.
-    IoError(io::Error),
-    /// The format of file is not supported
-    Format(String),
-    /// A decoding error, file is not a image, unsupported color space, etc.
-    Parsing(String),
-}
-
-/// An error that occurred if configuration is invalid
-///
-/// # Examples
-/// ```
-/// # use rimage::{Config, ConfigError};
-/// let config = Config::build(&[], 1.1, rimage::OutputFormat::MozJpeg);
-/// match config {
-///    Ok(_) => println!("Config is valid"),
-///  Err(e) => println!("Error: {}", e),
-/// }
-/// ```
-///
-/// # Errors
-///
-/// - [`ConfigError::QualityOutOfBounds`] if quality is less than 0 or greater than 1
-/// - [`ConfigError::InputIsEmpty`] if input is empty
-/// - [`ConfigError::WidthIsZero`] if width is 0
-/// - [`ConfigError::HeightIsZero`] if height is 0
-/// - [`ConfigError::SizeIsZero`] if size is 0
-///
-/// [`ConfigError::QualityOutOfBounds`]: enum.ConfigError.html#variant.QualityOutOfBounds
-/// [`ConfigError::InputIsEmpty`]: enum.ConfigError.html#variant.InputIsEmpty
-/// [`ConfigError::WidthIsZero`]: enum.ConfigError.html#variant.WidthIsZero
-/// [`ConfigError::HeightIsZero`]: enum.ConfigError.html#variant.HeightIsZero
-/// [`ConfigError::SizeIsZero`]: enum.ConfigError.html#variant.SizeIsZero
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum ConfigError {
-    /// Quality is less than 0 or greater than 1
-    QualityOutOfBounds,
-    /// Width is 0
-    WidthIsZero,
-    /// Height is 0
-    HeightIsZero,
-    /// Size is 0
-    SizeIsZero,
-    /// Input is empty
-    InputIsEmpty,
-}
-
-/// Color space of image
-///
-/// # Examples
-/// ```
-/// # use rimage::ColorSpace;
-/// # use std::str::FromStr;
-/// let color_space = ColorSpace::from_str("rgb").unwrap();
-/// println!("Color Space: {}", color_space);
-/// ```
-///
-/// # Errors
-///
-/// - [`ColorSpace::from_str`] if color space is not supported
-///
-/// [`ColorSpace::from_str`]: enum.ColorSpace.html#method.from_str
-#[derive(Debug, PartialEq, Eq)]
-pub enum ColorSpace {
-    /// **R**ed/**G**reen/**B**lue
-    Rgb,
-    /// **R**ed/**G**reen/**B**lue/**A**lpha
-    Rgba,
-    /// **C**yan/**M**agenta/**Y**ellow/Blac**K**
-    Cmyk,
-    /// Indexed color palette
-    Indexed,
-    /// Grayscale
-    Grayscale,
-    /// Grayscale/Alpha
-    GrayscaleAlpha,
-}
-
-impl std::str::FromStr for ColorSpace {
-    type Err = Cow<'static, str>;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "rgb" => Ok(ColorSpace::Rgb),
-            "rgba" => Ok(ColorSpace::Rgba),
-            "cmyk" => Ok(ColorSpace::Cmyk),
-            "indexed" => Ok(ColorSpace::Indexed),
-            "grayscale" => Ok(ColorSpace::Grayscale),
-            "grayscale_alpha" => Ok(ColorSpace::GrayscaleAlpha),
-            _ => Err(format!("{} is not a valid color space", s).into()),
-        }
-    }
-}
-
-impl fmt::Display for ColorSpace {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ColorSpace::Rgb => write!(f, "rgb"),
-            ColorSpace::Rgba => write!(f, "rgba"),
-            ColorSpace::Cmyk => write!(f, "cmyk"),
-            ColorSpace::Indexed => write!(f, "indexed"),
-            ColorSpace::Grayscale => write!(f, "grayscale"),
-            ColorSpace::GrayscaleAlpha => write!(f, "grayscale_alpha"),
-        }
-    }
-}
-
-/// Bit depth of image per pixel
-///
-/// # Examples
-/// ```
-/// # use rimage::BitDepth;
-/// # use std::str::FromStr;
-/// let bit_depth = BitDepth::from_str("8").unwrap();
-/// println!("Bit Depth: {}", bit_depth);
-/// ```
-///
-/// # Errors
-///
-/// - [`BitDepth::from_str`] if bit depth is not supported
-///
-/// [`BitDepth::from_str`]: enum.BitDepth.html#method.from_str
-#[derive(Debug)]
-#[repr(u8)]
-pub enum BitDepth {
-    /// One bit per pixel
-    One = 1,
-    /// Two bits per pixel
-    Two = 2,
-    /// Four bits per pixel
-    Four = 4,
-    /// Eight bits per pixel
-    Eight = 8,
-    /// Sixteen bits per pixel
-    Sixteen = 16,
-}
-
-impl std::str::FromStr for BitDepth {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "1" => Ok(BitDepth::One),
-            "2" => Ok(BitDepth::Two),
-            "4" => Ok(BitDepth::Four),
-            "8" => Ok(BitDepth::Eight),
-            "16" => Ok(BitDepth::Sixteen),
-            _ => Err(format!("{} is not a valid bit depth", s)),
-        }
-    }
-}
-
-impl fmt::Display for BitDepth {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            BitDepth::One => write!(f, "1"),
-            BitDepth::Two => write!(f, "2"),
-            BitDepth::Four => write!(f, "4"),
-            BitDepth::Eight => write!(f, "8"),
-            BitDepth::Sixteen => write!(f, "16"),
-        }
-    }
 }
 
 impl<'a> Decoder<'a> {
@@ -413,6 +175,7 @@ impl<'a> Decoder<'a> {
     ///
     /// [`Decoder`]: struct.Decoder.html
     /// [`io::Error`]: https://doc.rust-lang.org/std/io/struct.Error.html
+    #[inline]
     pub fn build(path: &'a path::PathBuf) -> Result<Self, DecodingError> {
         let raw_data = fs::read(path)?;
 
@@ -461,7 +224,7 @@ impl<'a> Decoder<'a> {
         panic::catch_unwind(|| -> Result<ImageData, DecodingError> {
             let d = mozjpeg::Decompress::new_mem(&self.raw_data)?;
             let color_space = match d.color_space() {
-                mozjpeg::ColorSpace::JCS_GRAYSCALE => ColorSpace::Grayscale,
+                mozjpeg::ColorSpace::JCS_GRAYSCALE => ColorSpace::Gray,
                 mozjpeg::ColorSpace::JCS_CMYK => ColorSpace::Cmyk,
                 mozjpeg::ColorSpace::JCS_RGB => ColorSpace::Rgb,
                 _ => ColorSpace::Rgb,
@@ -472,20 +235,17 @@ impl<'a> Decoder<'a> {
                 mozjpeg::Format::CMYK(img) => img,
             };
 
-            let data = Cow::Borrowed(image
-                .read_scanlines_flat()
-                .ok_or(DecodingError::Parsing(
-                    "Cannot read jpeg scanlines".to_string(),
-                ))?
-                .as_slice());
+            let data = image.read_scanlines_flat().ok_or(DecodingError::Parsing(
+                "Cannot read jpeg scanlines".to_string(),
+            ))?;
 
-            Ok(ImageData {
+            Ok(ImageData::new(
                 color_space,
-                bit_depth: BitDepth::Eight,
-                width: image.width(),
-                height: image.height(),
+                BitDepth::Eight,
+                image.width() as usize,
+                image.height() as usize,
                 data,
-            })
+            ))
         })
         .unwrap_or(Err(DecodingError::Parsing(
             "Failed to decode jpeg".to_string(),
@@ -498,156 +258,21 @@ impl<'a> Decoder<'a> {
         let mut buf = vec![0; reader.output_buffer_size()];
         let info = reader.next_frame(&mut buf)?;
         let data = buf[..info.buffer_size()].into();
-        Ok(ImageData {
-            color_space: info.color_type.into(),
-            bit_depth: info.bit_depth.into(),
-            width: info.width as usize,
-            height: info.height as usize,
+        Ok(ImageData::new(
+            info.color_type.into(),
+            info.bit_depth.into(),
+            info.width as usize,
+            info.height as usize,
             data,
-        })
+        ))
     }
 }
 
-impl<'a> ImageData<'a> {
-    /// Returns size of image (Width, Height)
-    ///
-    /// # Examples
-    /// ```
-    /// # use rimage::{Decoder, ImageData};
-    /// # use std::path;
-    /// # let path = path::PathBuf::from("tests/files/basi0g01.jpg");
-    /// # let d = Decoder::build(&path).unwrap();
-    /// # let image = d.decode().unwrap();
-    /// let (width, height) = image.size();
-    /// ```
-    pub fn size(&self) -> (usize, usize) {
-        (self.width, self.height)
-    }
-    /// Returns a ref to color space of image [`ColorSpace`]
-    ///
-    /// # Examples
-    /// ```
-    /// # use rimage::{Decoder, ImageData, ColorSpace};
-    /// # use std::path;
-    /// # let path = path::PathBuf::from("tests/files/basi0g01.jpg");
-    /// # let d = Decoder::build(&path).unwrap();
-    /// # let image = d.decode().unwrap();
-    /// let color_space = image.color_space();
-    /// match color_space {
-    ///     ColorSpace::Grayscale => println!("Grayscale"),
-    ///     ColorSpace::Rgb => println!("RGB"),
-    ///     ColorSpace::Cmyk => println!("CMYK"),
-    ///     ColorSpace::Rgba => println!("RGBA"),
-    ///     ColorSpace::Indexed => println!("Indexed"),
-    ///     ColorSpace::GrayscaleAlpha => println!("Grayscale Alpha"),
-    /// }
-    /// ```
-    /// [`ColorSpace`]: enum.ColorSpace.html
-    pub fn color_space(&self) -> &ColorSpace {
-        &self.color_space
-    }
-    /// Returns a ref to array of bytes in image
-    ///
-    /// # Examples
-    /// ```
-    /// # use rimage::{Decoder, ImageData};
-    /// # use std::path;
-    /// # let path = path::PathBuf::from("tests/files/basi0g01.jpg");
-    /// # let d = Decoder::build(&path).unwrap();
-    /// # let image = d.decode().unwrap();
-    /// let data = image.data();
-    /// ```
-    pub fn data(&self) -> &[u8] {
-        &self.data
-    }
-    /// Returns a ref to bit depth of image [`BitDepth`]
-    ///
-    /// # Examples
-    /// ```
-    /// # use rimage::{Decoder, ImageData, BitDepth};
-    /// # use std::path;
-    /// # let path = path::PathBuf::from("tests/files/basi0g01.jpg");
-    /// # let d = Decoder::build(&path).unwrap();
-    /// # let image = d.decode().unwrap();
-    /// let bit_depth = image.bit_depth();
-    /// match bit_depth {
-    ///     BitDepth::One => println!("1 bit"),
-    ///     BitDepth::Two => println!("2 bits"),
-    ///     BitDepth::Four => println!("4 bits"),
-    ///     BitDepth::Eight => println!("8 bits"),
-    ///     BitDepth::Sixteen => println!("16 bits"),
-    /// }
-    /// ```
-    ///
-    /// [`BitDepth`]: enum.BitDepth.html
-    pub fn bit_depth(&self) -> &BitDepth {
-        &self.bit_depth
-    }
-}
-
-impl From<io::Error> for DecodingError {
-    fn from(err: io::Error) -> Self {
-        DecodingError::IoError(err)
-    }
-}
-
-impl From<png::DecodingError> for DecodingError {
-    fn from(err: png::DecodingError) -> Self {
-        match err {
-            png::DecodingError::IoError(io_err) => DecodingError::IoError(io_err),
-            png::DecodingError::Format(f_err) => DecodingError::Format(f_err.to_string()),
-            png::DecodingError::Parameter(p_err) => DecodingError::Parsing(p_err.to_string()),
-            png::DecodingError::LimitsExceeded => {
-                DecodingError::Parsing("Png limits exceeded".to_string())
-            }
-        }
-    }
-}
-
-impl From<png::ColorType> for ColorSpace {
-    fn from(color_type: png::ColorType) -> Self {
-        match color_type {
-            png::ColorType::Grayscale => ColorSpace::Grayscale,
-            png::ColorType::Rgb => ColorSpace::Rgb,
-            png::ColorType::Indexed => ColorSpace::Indexed,
-            png::ColorType::GrayscaleAlpha => ColorSpace::GrayscaleAlpha,
-            png::ColorType::Rgba => ColorSpace::Rgba,
-        }
-    }
-}
-
-impl From<png::BitDepth> for BitDepth {
-    fn from(bit_depth: png::BitDepth) -> Self {
-        match bit_depth {
-            png::BitDepth::One => BitDepth::One,
-            png::BitDepth::Two => BitDepth::Two,
-            png::BitDepth::Four => BitDepth::Four,
-            png::BitDepth::Eight => BitDepth::Eight,
-            png::BitDepth::Sixteen => BitDepth::Sixteen,
-        }
-    }
-}
-
-impl fmt::Display for DecodingError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            DecodingError::IoError(io_err) => write!(f, "IO Error: {}", io_err),
-            DecodingError::Format(fmt_err) => write!(f, "Format Error: {}", fmt_err),
-            DecodingError::Parsing(prs_err) => write!(f, "Parsing Error: {}", prs_err),
-        }
-    }
-}
-
-impl fmt::Display for ConfigError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ConfigError::QualityOutOfBounds => write!(f, "Quality is out of bounds"),
-            ConfigError::WidthIsZero => write!(f, "Width cannot be zero"),
-            ConfigError::HeightIsZero => write!(f, "Height cannot be zero"),
-            ConfigError::SizeIsZero => write!(f, "Size cannot be zero"),
-            ConfigError::InputIsEmpty => write!(f, "Input cannot be zero"),
-        }
-    }
+/// Encoder used to encode image data to file
+pub struct Encoder<'a> {
+    path: &'a path::PathBuf,
+    image_data: ImageData,
+    quality: f32,
 }
 
 #[cfg(test)]
@@ -655,37 +280,6 @@ mod tests {
     use regex::Regex;
 
     use super::*;
-
-    #[test]
-    fn display_decoder_error() {
-        assert_eq!(
-            DecodingError::IoError(io::Error::new(io::ErrorKind::NotFound, "path not found"))
-                .to_string(),
-            "IO Error: path not found"
-        );
-        assert_eq!(
-            DecodingError::Format("WebP not supported".to_string()).to_string(),
-            "Format Error: WebP not supported"
-        );
-    }
-
-    #[test]
-    fn display_config_error() {
-        assert_eq!(
-            ConfigError::QualityOutOfBounds.to_string(),
-            "Quality is out of bounds"
-        );
-        assert_eq!(ConfigError::WidthIsZero.to_string(), "Width cannot be zero");
-        assert_eq!(
-            ConfigError::HeightIsZero.to_string(),
-            "Height cannot be zero"
-        );
-        assert_eq!(ConfigError::SizeIsZero.to_string(), "Size cannot be zero");
-        assert_eq!(
-            ConfigError::InputIsEmpty.to_string(),
-            "Input cannot be zero"
-        )
-    }
 
     #[test]
     fn decode_grayscale() {
@@ -706,7 +300,7 @@ mod tests {
 
             println!("{path:?}");
 
-            assert_eq!(image.color_space(), &ColorSpace::Grayscale);
+            assert_eq!(image.color_space(), &ColorSpace::Gray);
             assert_ne!(image.data().len(), 0);
             assert_ne!(image.size(), (0, 0));
         })
@@ -729,7 +323,7 @@ mod tests {
         files.iter().for_each(|path| {
             let image = Decoder::build(path).unwrap().decode().unwrap();
 
-            assert_eq!(image.color_space(), &ColorSpace::GrayscaleAlpha);
+            assert_eq!(image.color_space(), &ColorSpace::GrayAlpha);
             assert_ne!(image.data().len(), 0);
             assert_ne!(image.size(), (0, 0));
         })
